`getImageUrl` now returns a null/empty image instead of crashing/js error.
When the `getImageUrl` function can't load an image we now catch the error and return a null image instead. Previously we didn't do any error handling, this caused certain boards on the welcomescreen not to appear. This is somewhat unfortunate since unless the user manually goes into the config file to fix the path to the image that is causing errors, they wont be able to open that board anymore.

Now however, the function no longer crashes when it can't load an image from disk, and instead returns a null image. This makes it so the boards will still be displayed for example, all be it with the default "no image found/failed loading image" icon from HTML.