Fixed an issue with pasting a card that has a label, or pasting a list that contains cards with a label into a different board
The issue was that the labels on a card are actually references to the label colors which are stored in board. So upon pasting a card with a label into a different board, the user would encounter errors related to the card not being able to find the referenced color in the new board.

This fix makes it so that any missing colors in the new board the card is being pasted in are created automatically